requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
Django==1.11
django-crispy-forms==1.7.2
django-pam==1.4.1
enum34==1.1.6
idna==2.7
ipaddress==1.0.22
pamela==0.3.0
paramiko==2.4.1
pexpect==4.6.0
Pillow==5.2.0
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.18
PyNaCl==1.2.1
python-pam==1.8.4
pytz==2018.5
six==1.11.0
